requirements.txt
streamlit
google-generativeai
firebase-admin
pandas
folium
streamlit-folium
pillow
python-dotenv